+++ src/Check/index.js
@@ -2,9 +2,9 @@ import React from 'react'
 import cx from 'classnames'
 import {default as PT} from 'prop-types'
 
-const Check = ({ type, children, extraClass, ...props }) => {
+const Check = ({ type, children, extraClass, variant, ...props }) => {
   return (
-    <label className={cx('ola_check', extraClass)}>
+    <label className={cx('ola_check', variant && `is-${variant}`, extraClass)}>
       <input type={type} className="ola_check-input" {...props} />
       <div className="ola_check-label">
         <div className="ola_check-label-content">{ children }</div>
@@ -14,6 +14,7 @@ const Check = ({ type, children, extraClass, ...props }) => {
 }
 
 Check.defaultProps = {
+  variant: null,
   type: 'checkbox',
   extraClass: null,
 }
@@ -21,6 +22,8 @@ Check.defaultProps = {
 Check.propTypes = {
   /** Check Type */
   type: PT.oneOf(['checkbox', 'radio']),
+  /** CheckArea variants */
+  variant: PT.oneOf(['area', 'option']),
   /** Extra className */
   extraClass: PT.string,
   /** Childen nodes */