+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Added styles for `<hr>`
+
 ## [0.2.2] - 2019-11-27
 ### Fixed
 - The publish of the previous version
@@ -147,6 +151,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#45]: https://github.com/marketgoo/Ola/issues/45
 [#47]: https://github.com/marketgoo/Ola/issues/47
 
+[Unreleased]: https://github.com/marketgoo/Ola/compare/v0.2.2...HEAD
 [0.2.2]: https://github.com/marketgoo/Ola/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/marketgoo/Ola/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/marketgoo/Ola/compare/v0.1.2...v0.2.0