+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Default styles for task content [#54]
 - Improved Check label styles
 - Improved CheckArea styles [#56]
+- Fixed `.ola_ly-pageCenter` aligment in narrow screens.
 
 ## [0.2.2] - 2019-11-27
 ### Fixed